+++ src/tRelatorio.c
@@ -8,14 +8,14 @@ void relatorio1(int *vet, int t){
     printf("\n");
 }
 
-void relatorio2(float tempo, long long int comp, long long int trocas){
+void relatorio2(float tempo, unsigned long int comp, unsigned long int trocas){
     printf("Estatísticas gerais:\n");
     printf("Tempo de CPU: 	    %.5f segundos.\n",tempo);
-    printf("Comparações:	    %llu\n",comp);
-    printf("Trocas:             %llu\n\n",trocas); 
+    printf("Comparações:	    %lu\n",comp);
+    printf("Trocas:             %lu\n\n",trocas); 
 }
 
-void relatorio3(char algoritmo, char* arquivo, int tam, int T, long long int comp, long long int trocas, float tempo){
+void relatorio3(char algoritmo, char* arquivo, int tam, int T, unsigned long int comp, unsigned long int trocas, float tempo){
     printf("[algoritmo\tarquivo\t\t\ttam.\tT(top)\tcomp.\ttrocas\ttempo(s)]\n");
     if(algoritmo == 's'){
         printf(" seleção\t");
@@ -28,5 +28,5 @@ void relatorio3(char algoritmo, char* arquivo, int tam, int T, long long int com
     }else if(algoritmo == 'h' || algoritmo == 'a'){
         printf(" heapsort\t");
     }
-    printf("%s\t%d\t%d\t%llu\t%llu\t%.5f\n\n", arquivo, tam, T, comp, trocas, tempo);
+    printf("%s\t%d\t%d\t%lu\t%lu\t%.5f\n\n", arquivo, tam, T, comp, trocas, tempo);
 }
